fixes for octave compatibility
git-svn-id: svn+ssh://svn.code.sf.net/p/eidors3d/code/trunk@2728 82006732-fb7e-4a22-9f28-86a212cce563

+++ eidors/algorithms/a_adler/mk_c2f_circ_mapping.m
@@ -248,7 +248,10 @@ function frac= contained_elem_pts(m_pts, xyr);
            (m_pts(:,2,:) - xyr(2)).^2 + ...% yc =
            (m_pts(:,3,:) - xyr(3)).^2;     % zc =
      inpts = inr < xyr(4)^2;
-     frac= mean( inpts ,3);
+
+%    frac= mean( inpts ,3);
+%    FIXME: Octave doesn't like to mean on logical
+     frac= mean( int8( inpts ) ,3);
      if sum(inpts(:))==0
          % TODO: This message is outdated
          eidors_msg(['mk_c2f_circ_mapping: Interpolation failed: increase ', ...
